moved `logger` to the `util` folder
