+++ lib/patterns/controller.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 
 import 'package:flutter/material.dart';
-import 'package:module_provider/classes/logger.dart';
+import 'package:module_provider/util/logger.dart';
 import 'package:module_provider/module_provider.dart';
 import 'package:useful_classes/useful_classes.dart';
